PN-75 Migrate wallet-related methods that need web3 to Blockchain module.
